translator_application_with_streamlit/streamlit_requirements.txt
streamlit==0.89.0
googletrans==4.0.0-rc1
python-docx==0.8.11
PyPDF2==1.26.0